packaging: Improving support of lastlog
Since january 2015, ssh was complaining that /var/log/lastlog
was missing.

This is probably because the file /var/log/lastlog that is
%ghost in platform/upstream/setup is not created (but faillog is
created) during image creation.

As this have negative inpact due to error messages,
we decide to remove the support of lastlog in ssh in favor of the
support of lastlog in pam_lastlog.

This is good because management of lastlog is now made in a unique
place.

Change-Id: I22e31aa3c2f092323b9d692a273e938c1b7f9a85
Signed-off-by: José Bollo <jose.bollo@open.eurogiciel.org>